pyproject.toml
[tool.poetry]
name = "electricdreams"
version = "0.0.1"
description = "Helper module to interact with ChatGPT and Stable Diffusion. Just for DIY small projects."
authors = ["AlienMind <falienmind@gmail.com>"]
repository = "https://github.com/alienmind/electricdreams"
documentation = "https://alienmind.github.io/electricdreams/"
readme = "README.md"
packages = [
  {include = "electricdreams", from = "src"},
  {include = "electricdreams-cli", from = "src"},
  {include = "electricdreams-api", from = "src"}
]

[tool.poetry.scripts]
electricdreams-cli = "electricdreams-cli.cli:main"  
electricdreams-api = "electricdreams-api.api:run"  

[tool.poetry.dependencies]
python = ">=3.8,<=3.11"
python-dotenv = "^0.21.0"
openai = "^0.25.0"
diffusers = "0.8.0"
transformers = "^4.25.1"
scipy = "^1.9.3"
ftfy = "^6.1.1"
accelerate = "^0.15.0"
torch = "^1.13.0"
fastapi = "^0.88.0"
argparse = "^1.4.0"
uvicorn = "^0.20.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.2.0"
pytest-cov = "^4.0.0"
deptry = "^0.6.4"
mypy = "^0.981"
pre-commit = "^2.20.0"
tox = "^3.25.1"

[tool.poetry.group.docs.dependencies]
mkdocs = "^1.4.2"
mkdocs-material = "^8.5.10"
mkdocstrings = {extras = ["python"], version = "^0.19.0"}

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 120
target-version = ['py37']
preview = true

[tool.isort]
profile = "black"

[tool.mypy]
files = ["src/electricdreams"]
disallow_untyped_defs = "True"
disallow_any_unimported = "True"
no_implicit_optional = "True"
check_untyped_defs = "True"
warn_return_any = "True"
warn_unused_ignores = "True"
show_error_codes = "True"

[tool.coverage.report]
skip_empty = true

[tool.coverage.run]
branch = true
source = ["src/electricdreams"]

